description実装

+++ cogs/util.py
@@ -31,23 +31,12 @@ class util_cog(extcommands.Cog):
 
         if "twitter.com" in message.content or "x.com" in message.content:
             res = send_vx_twitter(message)
-            await message.channel.send(res)
-            await message.delete()
+            if res is not None:
+                await message.channel.send(res)
+                await message.delete()
 
 
-    @commands.application_command(guild_ids=gids)
-    async def hello(self, ctx, *, member: discord.Member = None):
-        """Says hello"""
-        member = member or ctx.author
-
-        if self._last_member is None or self._last_member.id != member.id:
-            await ctx.respond(f'Hello {member.name}~')
-        else:
-            await ctx.respond(f'Hello {member.name}... This feels familiar.')
-        self._last_member = member
-
-
-    @commands.application_command(guild_ids=gids)
+    @commands.application_command(guild_ids=gids, description="テスト用の音声ファイルを再生します")
     async def sound_test(self, ctx):
         if ctx.voice_client is None:
             await join_to_authors_channel(ctx)
@@ -58,12 +47,12 @@ class util_cog(extcommands.Cog):
         await ctx.respond("Playing test music file!")
 
 
-    @commands.application_command(guild_ids=gids)
-    async def akeome(self, ctx, num: int):
+    @commands.application_command(guild_ids=gids, description="あけおめを指定回数送信します")
+    async def akeome(self, ctx, num: discord.Option(int, "あけおめの回数") = 1):
         for i in range(num):
             await ctx.respond("あけおめ ( 'ω')")
 
-    @commands.application_command(guild_ids=gids)
+    @commands.application_command(guild_ids=gids, description="ぶどう先生からのありがたいメッセージを再生します🍇")
     async def budo(self, ctx):
         if ctx.voice_client is None:
             await join_to_authors_channel(ctx)
@@ -74,8 +63,8 @@ class util_cog(extcommands.Cog):
         ctx.voice_client.play(source)
 
 
-    @commands.application_command(guild_ids=gids)
-    async def meigen(self, ctx, index: int = -1, loop: int = 1):
+    @commands.application_command(guild_ids=gids, description="ぶどう先生から名言を賜ります🍇")
+    async def meigen(self, ctx, index: discord.Option(int, "名言の番号") = -1, loop: discord.Option(int, "ループ回数") = 1):
         for i in range(loop):
             if index > 0:
                 try:
